+++ test-tools/issuer-front-end/src/Main.tsx
@@ -88,7 +88,7 @@ function attributeInputPlaceHolder(details: AttributeDetails): string {
     if (details.type === 'date-time') {
         return '2023-08-30T06:22:46Z';
     }
-    if (details.type === 'number') {
+    if (details.type === 'number' || details.type === 'integer') {
         return '1234';
     }
     return 'myString';
@@ -181,7 +181,7 @@ function parseAttributesFromForm(
         } else if (obj.value !== undefined) {
             if (obj.type === 'string') {
                 attributes[obj.tag] = obj.value;
-            } else if (obj.type === 'number') {
+            } else if (obj.type === 'number' || obj.type == 'integer') {
                 attributes[obj.tag] = BigInt(obj.value);
             } else if (obj.type === 'date-time') {
                 const date = new Date(obj.value.trim());
@@ -193,16 +193,27 @@ function parseAttributesFromForm(
                 attributes[obj.tag] = date;
             } else {
                 setParsingError(
-                    `Attribute ${obj.tag} has type ${obj.type}. Only the types string/number and date-time are supported.`
+                    `Attribute ${obj.tag} has type ${obj.type}. Only the types string/number/integer and date-time are supported.`
                 );
                 throw new Error(
-                    `Attribute ${obj.tag} has type ${obj.type}. Only the types string/number and date-time are supported.`
+                    `Attribute ${obj.tag} has type ${obj.type}. Only the types string/number/integer and date-time are supported.`
                 );
             }
         }
     });
     return attributes;
 }
+
+// Convert a hex string to a byte array
+const hexToBytes = (hex: string) => {
+    const bytes = [];
+
+    for (let c = 0; c < hex.length; c += 2) {
+        bytes.push(parseInt(hex.substr(c, 2), 16));
+    }
+
+    return bytes;
+};
 
 export default function Main(props: WalletConnectionProps) {
     const { activeConnectorType, activeConnector, activeConnectorError, connectedAccounts, genesisHashes } = props;
@@ -215,7 +226,7 @@ export default function Main(props: WalletConnectionProps) {
     const [transactionError, setTransactionError] = useState('');
     const [userInputError2, setUserInputError2] = useState('');
 
-    const [auxiliaryData, setAuxiliaryData] = useState<number[]>([1, 2, 3]);
+    const [auxiliaryData, setAuxiliaryData] = useState('83fe0d');
 
     const [credentialRegistryContratIndex, setCredentialRegistryContratIndex] = useState<number | undefined>(0);
 
@@ -316,7 +327,7 @@ export default function Main(props: WalletConnectionProps) {
 
     const changeAuxiliaryDataHandler = useCallback((event: ChangeEvent) => {
         const target = event.target as HTMLTextAreaElement;
-        setAuxiliaryData(Array.from(JSON.parse(target.value)));
+        setAuxiliaryData(target.value);
     }, []);
 
     const changeSeedHandler = useCallback((event: ChangeEvent) => {
@@ -820,13 +831,13 @@ export default function Main(props: WalletConnectionProps) {
                                     onChange={changeCredentialMetaDataURLHandler}
                                 />
                                 <br />
-                                Add `AuxiliaryData`:
+                                Add `AuxiliaryData` (the hex string will be converted into bytes):
                                 <br />
                                 <input
                                     className="inputFieldStyle"
                                     id="auxiliaryDataTestCase4"
                                     type="text"
-                                    value={auxiliaryData.toString()}
+                                    value={auxiliaryData}
                                     onChange={changeAuxiliaryDataHandler}
                                 />
                                 <div className="switch-wrapper">
@@ -916,7 +927,7 @@ export default function Main(props: WalletConnectionProps) {
                                                     credentialMetaDataURL,
                                                     isHolderRevocable,
                                                     credentialRegistryContratIndex,
-                                                    auxiliaryData
+                                                    hexToBytes(auxiliaryData)
                                                 );
 
                                                 tx.then(setTxHash).catch((err: Error) =>
@@ -1064,13 +1075,13 @@ export default function Main(props: WalletConnectionProps) {
                                     onChange={changeReasonHandler}
                                 />
                                 <br />
-                                Add `AuxiliaryData`:
+                                Add `AuxiliaryData` (the hex string will be converted into bytes):
                                 <br />
                                 <input
                                     className="inputFieldStyle"
                                     id="auxiliaryDataTestCase6"
                                     type="text"
-                                    value={auxiliaryData.toString()}
+                                    value={auxiliaryData}
                                     onChange={changeAuxiliaryDataHandler}
                                 />
                                 <br />
@@ -1085,7 +1096,7 @@ export default function Main(props: WalletConnectionProps) {
                                             account,
                                             publicKey,
                                             credentialRegistryContratIndex,
-                                            auxiliaryData,
+                                            hexToBytes(auxiliaryData),
                                             reason
                                         );
 
@@ -1386,13 +1397,13 @@ export default function Main(props: WalletConnectionProps) {
                                     onChange={changeManualCredentialSchemaHandler}
                                 />
                                 <br />
-                                Add `AuxiliaryData`:
+                                Add `AuxiliaryData` (the hex string will be converted into bytes):
                                 <br />
                                 <input
                                     className="inputFieldStyle"
                                     id="auxiliaryDataTestCase11"
                                     type="text"
-                                    value={auxiliaryData.toString()}
+                                    value={auxiliaryData}
                                     onChange={changeAuxiliaryDataHandler}
                                 />
                                 <div className="switch-wrapper">
@@ -1550,7 +1561,7 @@ export default function Main(props: WalletConnectionProps) {
                                             credentialMetaDataURL,
                                             isHolderRevocable,
                                             credentialRegistryContratIndex,
-                                            auxiliaryData
+                                            hexToBytes(auxiliaryData)
                                         );
 
                                         tx.then(setTxHash).catch((err: Error) =>
@@ -1659,13 +1670,13 @@ export default function Main(props: WalletConnectionProps) {
                                     onChange={changeCredentialMetaDataURLHandler}
                                 />
                                 <br />
-                                Add `AuxiliaryData`:
+                                Add `AuxiliaryData` (the hex string will be converted into bytes):
                                 <br />
                                 <input
                                     className="inputFieldStyle"
                                     id="auxiliaryDataTestCase12"
                                     type="text"
-                                    value={auxiliaryData.toString()}
+                                    value={auxiliaryData}
                                     onChange={changeAuxiliaryDataHandler}
                                 />
                                 <div className="switch-wrapper">
@@ -1747,7 +1758,7 @@ export default function Main(props: WalletConnectionProps) {
                                                     credentialMetaDataURL,
                                                     isHolderRevocable,
                                                     credentialRegistryContratIndex,
-                                                    auxiliaryData
+                                                    hexToBytes(auxiliaryData)
                                                 );
 
                                                 tx.then(setTxHash).catch((err: Error) =>
@@ -1880,13 +1891,13 @@ export default function Main(props: WalletConnectionProps) {
                                     onChange={changeCredentialMetaDataURLHandler}
                                 />
                                 <br />
-                                Add `AuxiliaryData`:
+                                Add `AuxiliaryData` (the hex string will be converted into bytes):
                                 <br />
                                 <input
                                     className="inputFieldStyle"
                                     id="auxiliaryDataTestCase13"
                                     type="text"
-                                    value={auxiliaryData.toString()}
+                                    value={auxiliaryData}
                                     onChange={changeAuxiliaryDataHandler}
                                 />
                                 <div className="switch-wrapper">
@@ -1973,7 +1984,7 @@ export default function Main(props: WalletConnectionProps) {
                                                     credentialMetaDataURL,
                                                     isHolderRevocable,
                                                     credentialRegistryContratIndex,
-                                                    auxiliaryData
+                                                    hexToBytes(auxiliaryData)
                                                 );
 
                                                 tx.then(setTxHash).catch((err: Error) =>