+++ test-tools/issuer-front-end/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased changes
 
+## 1.0.9
+
+- Change to use a hexString for the auxiliary data.
+- Support 'integer' type in the schema.
+
 ## 1.0.8
 
 - Add support for date-time attributes.