+++ test-tools/proof-explorer/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased changes.
 
+## 1.0.3
+
+Support 'integer' type in the schema.
+
 ## 1.0.2
 
 Add support for date-time attributes.